backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-multipart
pytesseract
Pillow
scikit-learn
pandas
numpy
